dptb/requirements.txt
# Djando + util
Django~=4.2
pytils~=0.4.1
python-dotenv~=1.0.0

# Django productions
gunicorn~=20.1.0
whitenoise~=6.3.0

# Debagger
django_debug_toolbar>=4.0.0

# Bot & utils
beautifulsoup4~=4.11.1
dateparser~=1.1.5
python-dateutil~=2.8.2
python-telegram-bot~=13.15
timezonefinder~=6.2.0
tzlocal~=4.3

# detect language
langdetect~=1.0.9

# ChatGPT
openai~=0.27.5

# Distributed async tasks
celery~=5.2.7
celery[redis]
django-celery-beat~=2.5.0
django-redis~=5.2.0
redis~=4.5.4

# Common utilities
ipython>=8.7.0
